Move snippet building code
