+++ src/doc.ts
@@ -1,3 +1,5 @@
+import {workspace, SnippetString, WorkspaceConfiguration} from 'vscode';
+
 export class Doc {
     public params:Array<Param> = [];
     public return:string;
@@ -23,6 +25,68 @@ export class Doc {
                 this.params.push(new Param(param.type, param.name));
             });
         }
+    }
+
+    getConfig():WorkspaceConfiguration {
+        return workspace.getConfiguration('php-docblocker');
+    }
+
+    build():SnippetString {
+        let snippet = new SnippetString();
+        let extra = this.getConfig().get('extra');
+        let gap = !this.getConfig().get('gap');
+
+        let stop = 2;
+
+        snippet.appendText("/**");
+        snippet.appendText("\n * ");
+        snippet.appendVariable('1', this.message);
+
+        if (this.params && this.params.length) {
+            if (!gap) {
+                snippet.appendText("\n * ");
+                gap = true;
+            }
+            this.params.forEach(param => {
+                snippet.appendText("\n * @param ");
+                snippet.appendVariable(stop++ + '', param.type);
+                snippet.appendText(" ");
+                snippet.appendVariable(stop++ + '', param.name);
+            });
+        }
+
+        if (this.var) {
+            if (!gap) {
+                snippet.appendText("\n * ");
+                gap = true;
+            }
+            snippet.appendText("\n * @var ");
+            snippet.appendVariable(stop++ + '', this.var);
+        }
+
+        if (this.return) {
+            if (!gap) {
+                snippet.appendText("\n * ");
+                gap = true;
+            }
+            snippet.appendText("\n * @return ");
+            snippet.appendVariable(stop++ + '', this.return);
+        }
+
+        if (Array.isArray(extra) && extra.length > 0) {
+            if (!gap) {
+                snippet.appendText("\n * ");
+                gap = true;
+            }
+            for (var index = 0; index < extra.length; index++) {
+                var element = extra[index];
+                snippet.appendText("\n * " + element);
+            }
+        }
+
+        snippet.appendText("\n */");
+
+        return snippet;
     }
 }
